fix: session key validation error (#889)
* fix: session key validation error

When checking for covered calls in a session key, the control flow erroneously falls through when the scope signature is a hex string and the signature matches the call function selector. This causes an error in the next check when the code tries to pass in the scope signature to AbiItem.getSelector which throws an error "Unable to normalize signature." because it's not a valid human-readable function signature.

* Create unlucky-hotels-live.md

---------

Co-authored-by: jxom <7336481+jxom@users.noreply.github.com>